requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.31.0
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.4
openai==1.12.0
python-dateutil==2.9.0.post0 